:recycle: move slack events type definition to lib/slack
